Add files via upload
Scenario:

Consider a Chocolate Factory which produces various types of chocolates like milk chocolate, dark chocolate, white chocolate. The consumers can request for the required types of chocolates through the factory. However from the consumer's perspective they are completely unaware of who is creating this object. They just know that the Factory is providing them the required chocolates.

Problem Statement:

It is a standard practice that objects are created by calling the “new” keyword. Imagine a scenario where there are multiple cases in the client class and we call multiple new keywords for creating new objects.

if (selection.equalsIgnoreCase("white")) {
   return new WhiteChocolate();
  } else if (selection.equalsIgnoreCase("dark")) {
   return new DarkChocolate();
  }

If we now have to add organic chocolate or caramel chocolate we have to keep on modifying the client code and further add the new keyword. This creates a dependency on the client code and in turn makes it difficult to maintain.

Another problem is that the client application has to know how many types of concrete classes are available upfront. Later if we have to add another concrete class e.g. organic or semisweet or caramel then client code has to be changed and recompiled.

Input:

Select your chocolate?
White
Dark
Milk
Select your toppings
Sugar balls
Coconut powder
Nuts
Choco chips

Output:

Enjoy your <type> chocolate with <topping>